@@ -32,18 +32,21 @@
     "@reduxjs/toolkit": "^1.5.0",
     "ramda": "^0.28.0",
     "ramda-adjunct": "^3.0.0",
-    "react-redux": "^8.0.0",
     "serialize-error": "^10.0.0"
   },
   "peerDependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.0.0 || ^17.0.0",
+    "react-dom": "^18.0.0 || ^17.0.0",
+    "react-redux": "^8.0.0 || ^7.0.0"
   },
   "peerDependenciesMeta": {
     "react": {
       "optional": false
     },
     "react-dom": {
+      "optional": false
+    },
+    "react-redux": {
       "optional": false
     }
   },
