@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { System, SystemContext, useSystemComponent } from 'swagger-adjust';
 import { Provider } from 'react-redux';
 
@@ -22,15 +22,17 @@ const App = () => {
   return <TodoListLayout />;
 };
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = ReactDOM.createRoot(container);
+
+root.render(
   <React.StrictMode>
     <Provider store={store}>
       <SystemContext.Provider value={system.getSystem}>
         <App />
       </SystemContext.Provider>
     </Provider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
