@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import TextField from '@material-ui/core/TextField';
-import { makeStyles } from '@material-ui/core/styles';
+import TextField from '@mui/material/TextField';
+import { makeStyles } from '@mui/styles';
 import { useSystemActionCreatorBound } from 'swagger-adjust';
 
 const useStyles = makeStyles(() => ({
